requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
google-cloud-vision==3.8.0
Pillow==10.4.0
